Rabbitmq fixes (#43)
This changes the rabbitmq deployment to simplify it by not requiring extra plugins and service account/roles with permissions, etc. It leverages the predictable hostnames given by statefulset to form a cluster using the classic mode (considering replicatCount of stateful set - so single-node is permissible). It also updates the rabbitmq image version.