add concurrency restriction to pages publish job (#45)
This PR separates the helm build/publish into a separate job so that it can leverage the concurrency groups feature of gihub actions. Since the publication job checks out and updates the pages branch, there is a possible race condition when multiple builds are happening at the same time, resulting in all but one failing the push step. The concurrency group should resolve that.